// plugins/plugin2/Interfaces.ts
export interface Feature {
  runFeature: () => void;
  goToNextFeature: boolean;
}

export interface EventHandler {
  eventType: string;
  handleEvent: () => void;
}

export interface PipelineExecutor {
  features: Feature[];
  pipelineType: string;
  addFeature: (feature: Feature) => void;
}

export interface Plugin {
  eventHandlers: EventHandler[];
  pipelineExecutors: PipelineExecutor[];
  priority: number;
}